Combine interface for message and steps.
