+++ src/interfaces/MessageDisplay.ts
@@ -1,5 +1,3 @@
-import React, { ReactNode } from 'react';
-
 export interface Detail {
   id: number;
   header: string;
@@ -10,4 +8,13 @@ export interface MessageTypes {
   messSucc: boolean;
   messPassErr: boolean;
   messageDetails: Detail[];
+}
+
+export interface Steps {
+  id: number;
+  stepShow: boolean;
+  stepClass: string;
+  stepTitle: string;
+  stepDescription: string;
+  stepIcon: string;
 }
