--- src/interfaces/StepsInfo.ts
@@ -1,8 +1,0 @@
-export interface Steps {
-  id: number;
-  stepShow: boolean;
-  stepClass: string;
-  stepTitle: string;
-  stepDescription: string;
-  stepIcon: string;
-}
